pyproject.toml
[tool.poetry]
name = "news-mapping"
version = "0.1.0"
description = ""
authors = ["filopacio <filippo.pacinelli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
"yaml-1.3" = "^0.1.0"
groq = "^0.11.0"
pathlib = "^1.0.1"
matplotlib = "^3.9.2"
networkx = "^3.4.2"
pre-commit = "^4.0.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
serpapi = "^0.1.5"
google-search = "^1.1.1"
bs4 = "^0.0.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
